+++ R/padronizar_municipios.R
@@ -44,5 +44,9 @@ padronizar_municipios <- function(municipios) {
     combine = "or"
   )
 
-  padronizar_municipios_rs(enc2utf8(as.character(municipios)))
+  dedup <- indices_de_unicos(municipios)
+
+  padronizar_municipios_rs(
+    enc2utf8(as.character(dedup$unicos))
+  )[dedup$indices]
 }